Allow the user to specify which LaTeX compiler to use (#8)
We were using pdflatex as a default LaTeX compiler. This PR allows the user to specify any of the ones shipped with TeXLive.

This is particularly important because some applications require, e.g., Times New Roman font. In PDFLatex, it looks like it's basically impossible to guarantee this. In XeLaTeX and LuaLaTeX it seems to be possible. So having the option to switch between them can be useful.